move logic login reg current-user from user controller to auth controller
