+++ controllers/User.controller.js
@@ -1,5 +1,6 @@
 const db = require("../models");
 const user_model = db.User;
+const token_model = db.Token;
 
 const bcrypt = require("bcrypt");
 const jwt = require("jsonwebtoken");
@@ -16,138 +17,24 @@ const { errors } = require("pg-promise");
 
 
 
-const register = async (req, res) => {
-  const { error, value } = createUserSchema.validate(req.body);
-
-  if (error) {
-    return res.status(400).send({
-      message: "Validation error",
-      error: error.details,
-    });
-  }
-
-  try {
-    // Hash password with salt rounds 10
-    const hashedPassword = await bcrypt.hash(value.password, 10);
-
-    // Check email
-    const existingEmail = await user_model.findOne({
-      where: { email: value.email },
-    });
-
-    if (existingEmail) {
-      return res
-        .status(400)
-        .send({
-          message: "มีอีเมลล์นี้อยู่ในระบบอยู่แล้วกรุณากรอกอีเมลล์อื่น",
-        });
-    }
-
-    // Check username
-    const existingUsername = await user_model.findOne({
-      where: { username: value.username },
-    });
-
-    if (existingUsername) {
-      return res
-        .status(400)
-        .send({ message: "มี username นี้อยู่ในระบบอยู่แล้วกรุณาใช้ชื่ออื่น" });
-    }
-
-    // If both email and username are available, create the user
-    const user = await user_model.create({
-      email: value.email,
-      username: value.username,
-      password: hashedPassword,
-      title: value.title,
-      fname: value.fname,
-      lname: value.lname,
-      phone: value.phone,
-      role: value.role,
-      status: value.status,
-    });
-
-    return res.status(201).send({
-      message: "success"
-    });
-  } catch (err) {
-    console.log(err);
-    return res
-      .status(500)
-      .send({ message: "Server error", error: err.message });
-  }
-};
-
-const login = async (req, res) => {
-  const { error, value } = loginSchema.validate(req.body);
-
-  if (error) {
-    return res
-      .status(400)
-      .send({ message: "Validation error", error: error.details });
-  }
-
-  try {
-    const username = value.username;
-
-    const user = await user_model.findOne({
-      where: {
-        username: username,
-      },
-    });
-
-    if (!user) {
-      return res.status(404).send({ message: "ชื่อผู้ใช้ หรือ รหัสผ่านไม่ถูกต้อง" });
-    }
-
-    const passwordValit = await bcrypt.compare(value.password, user.password);
-
-    if (!passwordValit) {
-      return res.status(401).send({ message: "รหัสผ่านไม่ถูกต้อง" });
-    }
-
-    const token = jwt.sign(
-      {
-        id: user.id,
-        username: user.username,
-        role: user.role,
-        status: user.status,
-      },
-      secret,
-      { expiresIn: "24h" }
-    );
-
-    return res.status(200).send({
-      message: "Login success",
-      data: {
-        user: {
-          id: user.id,
-          username: user.username,
-          role: user.role,
-          status:user.status
-        },
-      },
-      Token: token,
-    });
-  } catch (err) {
-    console.log(err);
-    return res.status(500).send({ message: "Sever error", err: err.message });
-  }
-};
-
-const currentUser = async(req,res)=>{
+const loginHistory = async(req,res)=>{
   try{
-    const user = await user_model.findOne({
+    const results = await token_model.findAll({
       where:{
-        id: req.user.id
+        user_id: req.user.id
       },
-      attributes:{exclude:['password','phone']}
+      order:[["createdAt","DESC"]], //ล่าสุดมาก่อน
+      attributes:["createdAt","expiresAt"]
     })
 
-    return res.status(200).send(user)
+    if (results.length === 0) {
+      return res.status(404).send({ message: "ไม่พบประวัติการเข้าสู่ระบบ" });
+    }
 
-  }catch(e){
-    return res.status(500).send({message:'Sever error',errors:e})
+    return res.status(200).send({message:'success',results:results})
+
+  }catch(err){
+    return res.status(500).send({message:"Sever error",errors:err.message})
   }
 }
 
@@ -246,11 +133,10 @@ const approveUser = async(req,res)=>{
 }
 
 module.exports = { 
-  register,
-   login,currentUser,
    userList,
    findOneUser,
    deleteUser,
    findNonApprove,
-   approveUser
+   approveUser,
+   loginHistory
    };
