+++ router/auth/auth.route.js
@@ -6,7 +6,7 @@ const { auth,allowRole } = require('../../middleware/auth')
 
 router
     .post("/register",AuthController.register)
-    .post("/login", AuthController.loginUser)
-    .post("/current-user",auth, AuthController.currentUser)
+    .post("/login", AuthController.login)
+    .get("/current-user",auth, AuthController.currentUser)
 
 module.exports = router;
